Fixing Errors, Removing Windows Pollution from Rendering Module

+++ VanguardEngine/Source/Rendering/DescriptorHeap.cpp
@@ -1,12 +1,13 @@
 // Copyright (c) 2019-2020 Andrew Depke
 
 #include <Rendering/DescriptorHeap.h>
+#include <Rendering/Device.h>
 
 void DescriptorHeap::Initialize(RenderDevice& Device, D3D12_DESCRIPTOR_HEAP_TYPE Type, size_t Descriptors)
 {
 	D3D12_DESCRIPTOR_HEAP_DESC HeapDesc{};
 	HeapDesc.Type = Type;
-	HeapDesc.NumDescriptors = Descriptors;
+	HeapDesc.NumDescriptors = static_cast<UINT>(Descriptors);
 	HeapDesc.Flags = D3D12_DESCRIPTOR_HEAP_FLAG_NONE;  // #TODO: Shader visible?
 	HeapDesc.NodeMask = 0;
